# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SME Tax & Finance Assistant Algeria API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UpdateCompanyBodyActivityType =
  (typeof UpdateCompanyBodyActivityType)[keyof typeof UpdateCompanyBodyActivityType];

export const UpdateCompanyBodyActivityType = {
  commerce: "commerce",
  services: "services",
  production: "production",
  auto_entrepreneur: "auto_entrepreneur",
} as const;

export type UpdateCompanyBodyTaxRegime =
  (typeof UpdateCompanyBodyTaxRegime)[keyof typeof UpdateCompanyBodyTaxRegime];

export const UpdateCompanyBodyTaxRegime = {
  real: "real",
  simplified_real: "simplified_real",
  forfaitaire: "forfaitaire",
} as const;

export interface UpdateCompanyBody {
  companyName: string;
  activityType: UpdateCompanyBodyActivityType;
  /** @nullable */
  annualRevenue?: number | null;
  /** @nullable */
  employeeCount?: number | null;
  taxRegime: UpdateCompanyBodyTaxRegime;
  hasCnas?: boolean;
}

export interface Company {
  id: number;
  userId: number;
  companyName: string;
  activityType: string;
  /** @nullable */
  annualRevenue?: number | null;
  /** @nullable */
  employeeCount?: number | null;
  taxRegime: string;
  hasCnas: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CalendarEventStatus =
  (typeof CalendarEventStatus)[keyof typeof CalendarEventStatus];

export const CalendarEventStatus = {
  upcoming: "upcoming",
  due_today: "due_today",
  overdue: "overdue",
  completed: "completed",
} as const;

export type CalendarEventUrgencyLevel =
  (typeof CalendarEventUrgencyLevel)[keyof typeof CalendarEventUrgencyLevel];

export const CalendarEventUrgencyLevel = {
  normal: "normal",
  warning: "warning",
  critical: "critical",
  overdue: "overdue",
} as const;

export interface CalendarEvent {
  id: number;
  /** G50, G12, G12BIS, CNAS, IRG, IBS, G51, G4 */
  declarationType: string;
  regime: string;
  title: string;
  titleAr: string;
  description?: string;
  dueDate: string;
  daysUntilDue: number;
  status: CalendarEventStatus;
  urgencyLevel: CalendarEventUrgencyLevel;
}

export interface NewsItem {
  id: number;
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  category: string;
  publishedAt: string;
  createdAt: string;
}

export interface DashboardSummary {
  upcomingDeadlines: number;
  overdueCount: number;
  nextDeadline?: CalendarEvent;
  totalPenaltyRisk: number;
  acknowledgedReminders: number;
  pendingReminders: number;
  regimeLabel: string;
  recentNews: NewsItem[];
}

export type PenaltyCalculateBodyRegime =
  (typeof PenaltyCalculateBodyRegime)[keyof typeof PenaltyCalculateBodyRegime];

export const PenaltyCalculateBodyRegime = {
  real: "real",
  simplified_real: "simplified_real",
  forfaitaire: "forfaitaire",
} as const;

export interface PenaltyCalculateBody {
  /** e.g. G50, G12, G12BIS, CNAS, IBS, IRG */
  declarationType: string;
  regime: PenaltyCalculateBodyRegime;
  /** ISO date string of the legal deadline */
  dueDate: string;
  /** ISO date string of actual payment */
  paymentDate: string;
  /** Amount of tax owed (DZD). 0 if no rights. */
  taxAmount: number;
  /** Whether there are payment rights (droits de paiement) */
  hasPaymentRights: boolean;
}

export interface PenaltyBreakdownItem {
  label: string;
  /** @nullable */
  rate?: number | null;
  amount: number;
}

export interface PenaltyResult {
  delayMonths: number;
  penaltyRate: number;
  penaltyAmount: number;
  /** @nullable */
  fixedFine?: number | null;
  totalDue: number;
  breakdown: PenaltyBreakdownItem[];
  message: string;
  declarationType: string;
  regime: string;
}

export interface Penalty {
  id: number;
  userId: number;
  declarationType: string;
  regime: string;
  dueDate: string;
  paymentDate: string;
  taxAmount: number;
  penaltyRate: number;
  penaltyAmount: number;
  totalDue: number;
  createdAt: string;
}

export type ReminderReminderType =
  (typeof ReminderReminderType)[keyof typeof ReminderReminderType];

export const ReminderReminderType = {
  seven_days: "seven_days",
  three_days: "three_days",
  one_day: "one_day",
  due_day: "due_day",
  overdue: "overdue",
} as const;

export interface Reminder {
  id: number;
  userId: number;
  calendarEventId: number;
  reminderType: ReminderReminderType;
  message: string;
  messageAr: string;
  isAcknowledged: boolean;
  scheduledFor: string;
  createdAt: string;
  event: CalendarEvent;
}

export interface KnowledgeItem {
  id: number;
  question: string;
  questionAr: string;
  answer: string;
  answerAr: string;
  /** @nullable */
  tip?: string | null;
  /** @nullable */
  tipAr?: string | null;
  category: string;
  /** @nullable */
  regime?: string | null;
  createdAt: string;
}

export interface CreateNewsBody {
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  category: string;
  publishedAt?: string;
}

export interface TaxRule {
  id: number;
  regime: string;
  declarationType: string;
  legalDeadlineDescription: string;
  legalDeadlineDescriptionAr: string;
  /** JSON string of penalty schedule */
  penaltySchedule: string;
  /** @nullable */
  fixedFine?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  notesAr?: string | null;
  /** @nullable */
  sourceDocument?: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateRuleBody {
  legalDeadlineDescription?: string;
  legalDeadlineDescriptionAr?: string;
  penaltySchedule?: string;
  /** @nullable */
  fixedFine?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  notesAr?: string | null;
  isActive?: boolean;
}

export type UpdateAdminUserBodyRole =
  (typeof UpdateAdminUserBodyRole)[keyof typeof UpdateAdminUserBodyRole];

export const UpdateAdminUserBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export interface UpdateAdminUserBody {
  role?: UpdateAdminUserBodyRole;
  isActive?: boolean;
}

export interface RegimeCount {
  regime: string;
  count: number;
}

export interface AdminStats {
  totalUsers: number;
  activeUsers: number;
  totalRules: number;
  totalKnowledgeItems: number;
  totalNewsItems: number;
  usersByRegime: RegimeCount[];
}

export type ListCalendarEventsParams = {
  year?: number;
  month?: number;
};

export type ListKnowledgeItemsParams = {
  category?: string;
  regime?: string;
  q?: string;
};

export type ListRulesParams = {
  regime?: string;
  declarationType?: string;
};