# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SME Tax & Finance Assistant Algeria API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List tax rules
      parameters:
        - name: regime
          in: query
          required: false
          schema:
            type: string
        - name: declarationType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxRule"
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get a tax rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a tax rule (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRule"

  # ─── Admin ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TaxRule:
      type: object
      properties:
        id:
          type: integer
        regime:
          type: string
        declarationType:
          type: string
        legalDeadlineDescription:
          type: string
        legalDeadlineDescriptionAr:
          type: string
        penaltySchedule:
          type: string
          description: JSON string of penalty schedule
        fixedFine:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        notesAr:
          type: ["string", "null"]
        sourceDocument:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, regime, declarationType, legalDeadlineDescription, legalDeadlineDescriptionAr, penaltySchedule, isActive, createdAt, updatedAt]
    UpdateRuleBody:
      type: object
      properties:
        legalDeadlineDescription:
          type: string
        legalDeadlineDescriptionAr:
          type: string
        penaltySchedule:
          type: string
        fixedFine:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        notesAr:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ─── Admin ─────────────────────────────────────────────────────────────────
...